seed functionality added
now we can generate new maps based on a seed. this implementation is inefficient (it shuffles the array based on that seed each time perlin() is called, multiplied by each octave, multiplied by each tile on the map/pixel on a texture) but will work. we can shuffle one time if the entire function is converted to a struct where we can store the parameters for the function and keep them in separate instances for different purposes.