Update README.md
Remove shops and items from procedural generation description. Items are not decided by procedural generation, but instead drop from defeating enemies for a more rewarding gameplay experience. Shops are not part of our final goals and are not part of our development plan, so we would like to remove it from the description.